requirements_fyp.txt
annotated-types==0.7.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
eval_type_backport==0.2.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.29.1
imageio==2.36.1
importlib_resources==6.4.5
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
pillow==10.2.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==6.1.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
regex==2024.11.6
safetensors==0.5.3
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
transformers==4.49.0
triton==3.1.0
wandb==0.19.1
zipp==3.21.0
zstandard==0.23.0
einops==0.8.0
gmpy2==2.1.2
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
packaging==24.2
pycparser==2.22
PySocks==1.7.1
pytorch-lightning==2.5.0
PyYAML==6.0.2
requests==2.32.3
torchmetrics==1.5.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
